Removed old home, new home is todo-list
